+++ projects/core/src/remult3/RepositoryImplementation.ts
@@ -7,7 +7,8 @@ import { allEntities, Remult, isBackend, queryConfig as queryConfig, setControll
 import { AndFilter, rawFilterInfo, entityFilterToJson, Filter, FilterConsumer, OrFilter } from "../filter/filter-interfaces";
 import { Sort } from "../sort";
 import { v4 as uuid } from 'uuid';
-import cuid from 'cuid';
+import { createId } from '@paralleldrive/cuid2';
+
 
 
 
@@ -1770,10 +1771,10 @@ export class Fields {
     static cuid<entityType = any>(...options: (FieldOptions<entityType, string> | ((options: FieldOptions<entityType, string>, remult: Remult) => void))[]) {
         return Field(() => String, {
             allowApiUpdate: false,
-            defaultValue: () => cuid(),
+            defaultValue: () => createId(),
             saving: (_, r) => {
                 if (!r.value)
-                    r.value = cuid();
+                    r.value = createId();
             }
         }, ...options);
     }